refactor(sauna-map): split overloaded useVisitForm hook
- Extracted form state management into `useVisitFormState`
- Extracted API CRUD operations into `useVisitCrud`
- Updated `EditorContext.tsx` to use the separated hooks
- Split unit tests into `useVisitFormState.test.ts` and `useVisitCrud.test.ts`
